Short messages get truncated when using marker
Short messages, for example "abc123" would be shown as "abc1" if
mark_position=1 and mark_encrypted=^B^C12>^C^B.
As a simple demonstration, the following is sent and received using the
previsou plugin revision:

Sent                    Received
1                       1
12                      12
123                     123
1234                    1234
12345                   1234
123456                  1234
1234567                 1234
12345678                1234
123456789               123456789
1234567890              1234567890
12345678901             12345678901
123456789012            123456789012

To avoid this error, the decrypted string (with markers) can not replace
the string that was sent to the function and thus silently "corrput" the
input parameters to the function. To fix the error, use
signal_continue() with the new string as argument.
One could argue that signal_stop() followed by signal_emit() would be
the proper way to go to make sure that all plugins/scripts gets the
decrypted message, but that's another potential issue. The current
behavoir is preserved using signal_continue() and thus this is the path
chosen.